remove some jank
